+++ fe/user/src/utils/map.ts
@@ -1,7 +1,6 @@
 import { BASIC_COORDINATE } from '~/constants/COORDINATE';
 import { PIN_SVG, TIED_EIGHTH_NOTES_SVG } from '~/constants/MAP';
-import { Pin } from '~/types/api.types';
-import { Coordinate, CoordinateBoundary } from '~/types/map.types';
+import { Coordinate, CoordinateBoundary, Pin } from '~/types/map.types';
 
 export const fitBoundsToPins = (pins: Pin[], map: naver.maps.Map) => {
   if (pins.length === 0) {
